--- server/src/endpoints/endpoint.ts
@@ -1,35 +1,0 @@
-/*
-Copyright (c) Bryan Hughes <bryan@nebri.us>
-
-This file is part of Home Lights.
-
-Home Lights is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-Home Lights is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with Home Lights.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-import { AppState } from '../common/types';
-import { getLights } from '../db/lights';
-import { getPatterns } from '../db/patterns';
-import { getScenes } from '../db/scenes';
-import { getZones } from '../db/zones';
-import { getSystemState } from '../device';
-
-export function getAppState(): AppState {
-  return {
-    zones: getZones(),
-    scenes: getScenes(),
-    patterns: getPatterns(),
-    lights: getLights(),
-    systemState: getSystemState()
-  };
-}
